+++ assets/src/components/activities/multiple_choice/actions.ts
@@ -4,6 +4,7 @@ import { RichText, Hint as HintType, Choice } from '../types';
 import { Maybe } from 'tsmonad';
 import { toSimpleText } from 'data/content/text';
 import { Identifiable } from 'data/content/model';
+import { PostUndoable } from 'components/activities/types';
 
 export class MCActions {
   private static getById<T extends Identifiable>(slice: T[], id: string): Maybe<T> {
@@ -26,7 +27,7 @@ export class MCActions {
   }
 
   static addChoice() {
-    return (draftState: MultipleChoiceModelSchema) => {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
       const newChoice: Choice = fromText('');
       draftState.choices.push(newChoice);
       draftState.authoring.parts[0].responses.push(
@@ -36,13 +37,13 @@ export class MCActions {
   }
 
   static editChoice(id: string, content: RichText) {
-    return (draftState: MultipleChoiceModelSchema) => {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
       MCActions.getChoice(draftState, id).lift((choice) => (choice.content = content));
     };
   }
 
   static removeChoice(id: string) {
-    return (draftState: MultipleChoiceModelSchema) => {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
       draftState.choices = draftState.choices.filter((c) => c.id !== id);
       draftState.authoring.parts[0].responses = draftState.authoring.parts[0].responses.filter(
         (r) => r.rule !== `input like {${id}}`,
@@ -51,13 +52,13 @@ export class MCActions {
   }
 
   static editFeedback(id: string, content: RichText) {
-    return (draftState: MultipleChoiceModelSchema) => {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
       MCActions.getResponse(draftState, id).lift((r) => (r.feedback.content = content));
     };
   }
 
   static addHint() {
-    return (draftState: MultipleChoiceModelSchema) => {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
       const newHint: HintType = fromText('');
       // new hints are always cognitive hints. they should be inserted
       // right before the bottomOut hint at the end of the list
@@ -67,16 +68,25 @@ export class MCActions {
   }
 
   static editHint(id: string, content: RichText) {
-    return (draftState: MultipleChoiceModelSchema) => {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
       MCActions.getHint(draftState, id).lift((hint) => (hint.content = content));
     };
   }
 
-  static removeHint(id: string) {
-    return (draftState: MultipleChoiceModelSchema) => {
+  static removeHint(id: string, ) {
+    return (draftState: MultipleChoiceModelSchema, post: PostUndoable) => {
+      const hint = draftState.authoring.parts[0].hints.find((h) => h.id === id);
+      const index = draftState.authoring.parts[0].hints.findIndex((h) => h.id === id);
       draftState.authoring.parts[0].hints = draftState.authoring.parts[0].hints.filter(
         (h) => h.id !== id,
       );
+      if (hint !== undefined) {
+        post({
+          description: 'Removed a hint',
+          operations: [{ path: '$.authoring.parts[0].hints', index, item: JSON.parse(JSON.stringify(hint as any))}],
+          type: 'Undoable',
+        })
+      }
     };
   }
 }