+++ assets/src/components/activity/PartObjectives.tsx
@@ -1,17 +1,15 @@
-import * as Immutable from 'immutable';
 import React from 'react';
-import { ResourceId } from 'data/types';
 import { Objective } from 'data/content/objective';
 import { Objectives } from 'components/resource/Objectives';
-import { valueOr } from 'utils/common';
+import { ObjectiveMap } from 'data/content/activity';
 
 export type PartObjectivesProps = {
-  partIds: Immutable.List<string>;
-  objectives: Immutable.Map<string, Immutable.List<ResourceId>>;
-  allObjectives: Immutable.List<Objective>;
+  partIds: string[];
+  objectives: ObjectiveMap;
+  allObjectives: Objective[];
   editMode: boolean;
   projectSlug: string;
-  onEdit: (objectives: Immutable.Map<string, Immutable.List<ResourceId>>) => void;
+  onEdit: (objectives: ObjectiveMap) => void;
   onRegisterNewObjective: (objective: Objective) => void;
 };
 
@@ -25,24 +23,18 @@ export const PartObjectives = (props: PartObjectivesProps) => {
         <div className="card-title">Learning Objectives</div>
         <div className="d-flex flex-row align-items-baseline">
           <div className="flex-grow-1">
-            {partIds.size > 1 ? (
-              partIds.toArray().map((id) => (
+            {partIds.length > 1 ? (
+              partIds.map((id) => (
                 <div key={id} className="d-flex flex-row align-items-baseline">
                   <div>Part {id}</div>
                   <Objectives
                     editMode={editMode}
                     projectSlug={props.projectSlug}
-                    selected={Immutable.List<ResourceId>(
-                      valueOr(objectives.get(id), Immutable.List<ResourceId>()),
-                    )}
+                    selected={objectives[id]}
                     objectives={allObjectives}
                     onRegisterNewObjective={onRegisterNewObjective}
                     onEdit={(objectives) =>
-                      onEdit(
-                        Immutable.Map<string, Immutable.List<ResourceId>>({
-                          [id]: objectives,
-                        } as any),
-                      )
+                      onEdit(Object.assign({}, props.objectives, { [id]: objectives }))
                     }
                   />
                 </div>
@@ -52,17 +44,11 @@ export const PartObjectives = (props: PartObjectivesProps) => {
                 <Objectives
                   editMode={editMode}
                   projectSlug={props.projectSlug}
-                  selected={Immutable.List<ResourceId>(
-                    valueOr(objectives.get(partIds.first()), Immutable.List<ResourceId>()),
-                  )}
+                  selected={objectives[partIds[0]]}
                   objectives={allObjectives}
                   onRegisterNewObjective={onRegisterNewObjective}
                   onEdit={(objectives) =>
-                    onEdit(
-                      Immutable.Map<string, Immutable.List<ResourceId>>({
-                        [partIds.first() as string]: objectives,
-                      } as any),
-                    )
+                    onEdit(Object.assign({}, props.objectives, { [partIds[0]]: objectives }))
                   }
                 />
               </div>