+++ assets/src/components/resource/editors/createEditor.tsx
@@ -1,16 +1,19 @@
 // Factory for creating top level editors, for things like structured
 
 import React from 'react';
-import { ResourceContent, ResourceType, Activity } from 'data/content/resource';
+import { ResourceContent, ResourceContext, ResourceType } from 'data/content/resource';
 import { StructuredContentEditor } from 'components/content/StructuredContentEditor';
-import { EditorDesc, ActivityEditorMap } from 'data/content/editors';
+import { EditorDesc } from 'data/content/editors';
 import { ContentBlock } from './ContentBlock';
 import { ActivityBlock } from './ActivityBlock';
 import { getToolbarForResourceType } from '../../editing/toolbars/insertion/items';
 import { UnsupportedActivity } from '../UnsupportedActivity';
 import * as Immutable from 'immutable';
-import { TestModeHandler, defaultState } from '../TestModeHandler';
-import { valueOr } from 'utils/common';
+import { defaultState } from '../TestModeHandler';
+import { ActivityEditContext } from 'data/content/activity';
+import { InlineActivityEditor, EditorUpdate } from 'components/activity/InlineActivityEditor';
+import { Objective } from 'data/content/objective';
+import { Undoable } from 'components/activities/types';
 
 const unsupported: EditorDesc = {
   deliveryElement: UnsupportedActivity,
@@ -25,17 +28,21 @@ const unsupported: EditorDesc = {
 
 // content or referenced activities
 export const createEditor = (
+  resourceContext: ResourceContext,
   content: ResourceContent,
   index: number,
-  activities: Immutable.Map<string, Activity>,
-  editorMap: ActivityEditorMap,
+  activities: Immutable.Map<string, ActivityEditContext>,
   editMode: boolean,
   resourceSlug: string,
   projectSlug: string,
   graded: boolean,
   objectivesMap: any,
   editorProps: any,
+  allObjectives: Objective[],
   onEdit: (content: ResourceContent) => void,
+  onActivityEdit: (key: string, update: EditorUpdate) => void,
+  onPostUndoable: (key: string, undoable: Undoable) => void,
+  onRegisterNewObjective: (o: Objective) => void,
 ): JSX.Element => {
   if (content.type === 'content') {
     return (
@@ -57,15 +64,7 @@ export const createEditor = (
   const activity = activities.get(content.activitySlug);
 
   if (activity !== undefined) {
-    const editor = editorMap[activity.typeSlug] ? editorMap[activity.typeSlug] : unsupported;
-
     const previewText = activity.model.authoring?.previewText;
-
-    // Test mode is supported by giving the delivery component a transformed
-    // instance of the activity model.  Recognizing that we are in an editing mode
-    // we make this robust to problems with transformation so we fallback to the raw
-    // model if the transformed model is null (which results from failure to transform)
-    const model = valueOr(activity.transformed, activity.model);
 
     const slugsAsKeys = Object.keys(activity.objectives).reduce((map: any, key) => {
       (activity.objectives as any)[key as any].forEach((slug: string) => {
@@ -77,25 +76,40 @@ export const createEditor = (
     const objectives = Object.keys(slugsAsKeys).map((slug) => objectivesMap[slug]);
 
     const props = {
-      model: JSON.stringify(model),
+      model: activity.model,
       activitySlug: activity.activitySlug,
-      state: JSON.stringify(defaultState(model)),
+      state: defaultState(activity.model),
+      typeSlug: activity.typeSlug,
+      editMode: editMode,
       graded: false,
+      projectSlug: projectSlug,
+      resourceSlug: resourceSlug,
+      resourceId: resourceContext.resourceId,
+      resourceTitle: resourceContext.title,
+      authoringScript: activity.authoringScript,
+      authoringElement: activity.authoringElement,
+      friendlyName: activity.friendlyName,
+      description: activity.description,
+      objectives: activity.objectives,
+      allObjectives: allObjectives,
+      activityId: activity.activityId,
+      title: activity.title,
+      onEdit: (update: EditorUpdate) => onActivityEdit(activity.activitySlug, update),
+      onPostUndoable: (undoable: Undoable) => onPostUndoable(activity.activitySlug, undoable),
+      onRegisterNewObjective,
     };
 
     return (
       <ActivityBlock
         {...editorProps}
         contentItem={content}
-        label={editor.friendlyName}
+        label={activity.friendlyName}
         projectSlug={projectSlug}
         resourceSlug={resourceSlug}
         objectives={objectives}
         previewText={previewText}
       >
-        <TestModeHandler model={model}>
-          {React.createElement(editor.deliveryElement, props as any)}
-        </TestModeHandler>
+        <InlineActivityEditor {...props} />
       </ActivityBlock>
     );
   }