+++ assets/src/components/resource/editors/dragndrop/handlers/focus.ts
@@ -1,20 +1,20 @@
 import * as Immutable from 'immutable';
-import { ResourceContent, Activity } from 'data/content/resource';
+import { ResourceContent } from 'data/content/resource';
 import { ActivityEditorMap } from 'data/content/editors';
 import { getContentDescription } from 'data/content/utils';
-import { getFriendlyName } from '../utils';
+import { ActivityEditContext } from 'data/content/activity';
 
 export const focusHandler = (
   setAssistive: (s: string) => void,
   content: Immutable.OrderedMap<string, ResourceContent>,
   editorMap: ActivityEditorMap,
-  activities: Immutable.Map<string, Activity>,
+  activities: Immutable.Map<string, ActivityEditContext>,
 ) => (key: string) => {
   const item = content.get(key) as ResourceContent;
   const desc =
     item.type === 'content'
       ? getContentDescription(item)
-      : getFriendlyName(item, editorMap, activities);
+      : activities.get(item.activitySlug)?.friendlyName;
 
   const index = content.keySeq().findIndex((k) => k === key);
   setAssistive(`Listbox. ${index + 1} of ${content.size}. ${desc}.`);